update to 1.20.0.10

+++ BuildTools/macosx/install_name.py
@@ -4,8 +4,9 @@ from shutil import copyfile
 
 framework_path = sys.argv[1] #e.g. '/Users/xun/Github/geoda/BuildTools/macosx/build/GeoDa.app/Contents/Frameworks'
 id = sys.argv[2]
+codesign_only = bool(sys.argv[3]) if len(sys.argv) > 3  else False
 
-print(id)
+print(id, codesign_only)
 
 def ProcessDependency(dir_path, dylib_name):
     dylib_path = dir_path + '/' + dylib_name
@@ -14,22 +15,23 @@ def ProcessDependency(dir_path, dylib_name):
     for item in items:
         item = item.strip().split(" ")[0]
 
-        # workaround for gdal 3.3.3 that @rpath/libgeos.3.10.1.dylib was used instead of dir path
+        # workaround for gdal 3.3.3 that @rpath/libgeos.3.10.2.dylib was used instead of dir path
         copyitem = item
-        if item == '@rpath/libgeos.3.10.1.dylib':
+        if item == '@rpath/libgeos.3.10.2.dylib':
             copyitem = '/usr/local/opt/geos/lib/libgeos.dylib'
 
-        if item.startswith('/usr/lib') == False and item.startswith('/System') == False and item.startswith('@executable_path/')==False:
+        if item.startswith('/usr/lib') == False and item.startswith('/System') == False and (codesign_only or item.startswith('@executable_path/')==False):
             print("Process:", item)
             file_name = os.path.basename(item)
             # Copy the dylib to Frameworks if needed
             dest = dir_path + '/' + file_name
-            if os.path.exists(dest) == False:
+            if os.path.exists(dest) == False and not codesign_only:
                 copyfile(copyitem, dest, follow_symlinks=True)
             # install_name_tool current item
             new_path = "@executable_path/../Frameworks/{}".format(file_name)
             cmd = "install_name_tool -change \"{}\" \"{}\" {}".format(item, new_path, dylib_path)
-            os.system(cmd)
+            if not codesign_only:
+                os.system(cmd)
             # process item
             ProcessDependency(dir_path, file_name)
     print("codesign {}", dylib_path)
@@ -39,4 +41,4 @@ def ProcessDependency(dir_path, dylib_name):
 
 ProcessDependency(framework_path, "libwx_osx_cocoau_gl-3.1.dylib")
 ProcessDependency(framework_path, "libwx_osx_cocoau-3.1.dylib")
-ProcessDependency(framework_path, "libgdal.29.dylib")
+ProcessDependency(framework_path, "libgdal.30.dylib")
